+++ CHANGELOG.md
@@ -6,6 +6,18 @@ For a brief few notes on what's Inane Class check out the [InaneClasses Wiki](ht
 
 Check out the [README](README.md) for installation stuff.
 
+## 0.26.0 - 2022-01-21
+
+- Config\Options:
+    - fix var_export empty of data
+- Debug\Dumper:
+    - variable name used (when able) if no label supplied
+    - label now shows type
+- File\FileInfo AND Http\Client
+    - fixes for php 8.1
+- Data\ValueType
+    - New enum for MySQL => php types
+
 ## 0.25.0 - 2021-12-07
 
 - Dumper        : minor fixes